# iterate through changed links only in UpdateSim