@adlarkin 2021-03-10
Note to self: some of the caching that was done in #656 may no longer be needed (I will look into this). #669 has already been removed in this PR since https://github.com/ignitionrobotics/ign-physics/pull/223 gives us only the links that have moved in a physics step