@chapulina 2021-03-15
> We have been trying not to require features in a released version

+1, we should be mindful of not introducing new required APIs that may break existing 3rd party physics engines.

> possible workaround to maintain backward compatibility

That sounds like a good idea. 

Another thing to consider is working on this feature on top of Edifice so that we're free to change requirements and implement the feature in the ideal way. Once that's settled, we can think of how to backport it.