@adlarkin 2021-03-15
Thanks for the comments, @azeey and @chapulina. I am going to update this PR and https://github.com/ignitionrobotics/ign-physics/pull/223 to target Edifice, and then we can determine the next best steps.