@adlarkin 2021-03-16
This is now ready for review, and targets Edifice (`main` branch).

As I mentioned in the PR description, this change eliminates the need for #669, which is something to keep in mind when we make the next forward-port.

We'll need to make a minor release for `ign-math6` so that ignitionrobotics/ign-math#196 can be used.