@adlarkin 2021-03-22
> As I mentioned in the PR description, this change eliminates the need for #669, which is something to keep in mind when we make the next forward-port.

I just updated this branch to include the changes in #695, and resolved conflicts to use the changes in this PR instead of the changes in #669: https://github.com/ignitionrobotics/ign-gazebo/pull/678/commits/1074b5a5cfddc2c8edb3f2975174c8f1b3105f8a