@adlarkin 2021-03-23
> LGTM!

I'm making one last change that will make this new physics output optional (if poses aren't a part of `ignition::physics::ForwardStep::Output`, then the physics system will iterate through all of the links in `UpdateSim` instead of through the links in `ignition::physics::ForwardStep::Output`). Once that change is in, I believe this can be merged.