@adlarkin 2021-03-23
> I'm making one last change that will make this new physics output optional

@azeey this new behavior is now optional: https://github.com/ignitionrobotics/ign-gazebo/pull/678/commits/040b4a16fe162c167852c519895fac01543c1876

In order to achieve optional behavior, this PR now requires https://github.com/ignitionrobotics/ign-gazebo/pull/678/commits/040b4a16fe162c167852c519895fac01543c1876.