@adlarkin 2021-03-25
Conflicts have been resolved in https://github.com/ignitionrobotics/ign-gazebo/pull/678/commits/dca0f6686cf6e7b7c0173a9f54bd58a414f04dc8, but it looks like there's a new test failure that I need to resolve (`INTEGRATION_wind_effects`). Using physics engine output to determine which links changed is still optional behavior (if a physics engine doesn't write output, the `ECM` is used like we've been doing up until this point).

@azeey, I had to make a few changes to the physics system when resolving merge conflicts in order to support the work you did in #685. The main difference that was made was including a `modelWorldPoses` cache since the frame data that is used in `UpdateSim` only contains links that have changed instead of all links (which means that we don't update every model's pose every time, so we should cache model poses to avoid losing pose information for models that aren't updated). The merge commit (https://github.com/ignitionrobotics/ign-gazebo/pull/678/commits/dca0f6686cf6e7b7c0173a9f54bd58a414f04dc8) shows how I resolved the conflicts and how I made use of this `modelWorldPoses` cache, and as we discussed offline, I made some optimizations to the model pose update `ecm.each` call in https://github.com/ignitionrobotics/ign-gazebo/pull/678/commits/25719238fee10560800cbbe81a97ef3db11f0669.

To give more insight into the benefits of https://github.com/ignitionrobotics/ign-gazebo/pull/678/commits/25719238fee10560800cbbe81a97ef3db11f0669, here are the profiler results before this commit using the same testing methodology described in the PR description. As we can see, the `ecm.Each` model loop takes `2.5ms`:

![ecm_each_original](https://user-images.githubusercontent.com/42042756/112408937-00988500-8cef-11eb-8d1b-e0068abf1d1c.png)

After applying the changes made in https://github.com/ignitionrobotics/ign-gazebo/pull/678/commits/25719238fee10560800cbbe81a97ef3db11f0669, the `ecm.Each` model loop takes `0.79ms`:

![modified_ecm_each](https://user-images.githubusercontent.com/42042756/112409165-5b31e100-8cef-11eb-824a-3b7164e5fe3b.png)

It's interesting to see how much faster the `ecm.Each` call can be when it is used with fewer components (in this case, we only use the components we absolutely need with `ecm.Each`, and then query other components later on if we need them). Perhaps optimizing `ecm.Each` is something I can look into as I continue to work on simulation performance updates.

---

**Testing the optional behavior**

To revert to the old behavior of using the `ECM` to go through all links for updates instead of using physics output for updating links, make the following changes in `ign-physics` (this is assuming you're on commit `b10a4c3` in `ign-physics`):
```diff
diff --git a/dartsim/src/SimulationFeatures.cc b/dartsim/src/SimulationFeatures.cc
index 2c4ef89..e7aa863 100644
--- a/dartsim/src/SimulationFeatures.cc
+++ b/dartsim/src/SimulationFeatures.cc
@@ -59,7 +59,7 @@ void SimulationFeatures::WorldForwardStep(
 
   // TODO(MXG): Parse input
   world->step();
-  this->Write(_h.Get<ChangedWorldPoses>());
+  //this->Write(_h.Get<ChangedWorldPoses>());
   // TODO(MXG): Fill in state
 }
 
diff --git a/tpe/plugin/src/SimulationFeatures.cc b/tpe/plugin/src/SimulationFeatures.cc
index 30db7b4..10992c3 100644
--- a/tpe/plugin/src/SimulationFeatures.cc
+++ b/tpe/plugin/src/SimulationFeatures.cc
@@ -62,7 +62,7 @@ void SimulationFeatures::WorldForwardStep(
     }
   }
   world->Step();
-  this->Write(_h.Get<ChangedWorldPoses>());
+  //this->Write(_h.Get<ChangedWorldPoses>());
 }
```

You can also add some print statements inside the `if`...`else` block of the `PhysicsPrivate::ChangedLinks` method in the physics system to get further validation on which method is being used for recording link updates.